requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
django-redis==5.4.0
django-environ==0.11.2
requests==2.31.0
django-crispy-forms
crispy-bootstrap5
django-ses==3.5.2
web3==6.11.4
coverage==7.3.2
